# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the aztec P2P Explorer API server.",
        "title": "Aztec P2P Explorer API",
        "contact": {
            "name": "aztec-p2p-explorer",
            "url": "http://github.com/NethermindEth/aztec-p2p-explorer"
        },
        "version": "1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/peers/map": {
            "get": {
                "description": "Get geographical clusters of peers with their locations and counts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peers"
                ],
                "summary": "Get peer clusters for map visualisation",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Continent names",
                        "name": "continents",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Continent codes",
                        "name": "continents_code",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Country names",
                        "name": "countries",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Country iso codes",
                        "name": "countries_codes",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "City names",
                        "name": "cities",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Names of Autonomous System Organisations",
                        "name": "as_names",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Numbers of Autonomous System Organisations",
                        "name": "as_numbers",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Client names (e.g. aztec-node)",
                        "name": "clients",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Sync status",
                        "name": "synced",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include peers from latest crawl only (default: true)",
                        "name": "latest",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PeersMapResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.APIError": {
            "type": "object",
            "properties": {
                "message": {},
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "server.Cluster": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "continent": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "peer_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "server.PeersMapResponse": {
            "type": "object",
            "properties": {
                "active_peers": {
                    "type": "integer"
                },
                "clusters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.Cluster"
                    }
                },
                "total_peers": {
                    "type": "integer"
                }
            }
        },
...
    }
}